# apps/backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: ""
      tags: &a1
        - auth
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenDto"
      responses:
        "201":
          description: ""
      tags: *a1
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a1
      security:
        - bearer: []
...
info:
  title: Wellness Package Management API
  description: Admin and mobile API endpoints
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginDto:
      type: object
      properties: {}
    RefreshTokenDto:
      type: object
      properties: {}
...